class - Separando componentes
